requirements.txt
numpy
opencv-python
deepface
pyserial
playsound==1.2.2
rpi_ws281x
adafruit-circuitpython-neopixel
board
